requirements.txt
numpy
pandas
nltk
spacy
swifter
faiss-cpu
sentence-transformers
transformers
datasets
tqdm
torch